+++ .env
@@ -13,3 +13,7 @@ POSTGRES_PORT=${POSTGRES_PORT:-5432}
 POSTGRES_DB=${POSTGRES_DB:-sadfavbotdb}
 POSTGRES_USER=${POSTGRES_USER:-sadfavbot}
 POSTGRES_PASSWORD=${POSTGRES_PASSWORD:-usrpwd4sfb}
+
+WEBHOOK_HOST=${WEBHOOK_HOST}
+WEBHOOK_PORT=${WEBHOOK_PORT:-443}
+WEBHOOK_PATH=${WEBHOOK_PATH:-SadFavBot}
