+++ docker-compose.yaml
@@ -18,8 +18,12 @@ services:
       - POSTGRES_DB
       - POSTGRES_USER
       - POSTGRES_PASSWORD
+      - WEBHOOK_HOST
+      - WEBHOOK_PORT
+      - WEBHOOK_PATH
     expose:
       - ${METRICS_PORT}
+      - ${WEBHOOK_PORT}
     depends_on:
       - postgres
       - redis
